Enhanced parsing and conversion functions for Insert queries
Fixed a bug where insert queries that specified a schema name could not be parsed.
Added a function to convert Select query to Values query (TryToValuesQuery).
Added a query to convert a Values query to a Select query without a table (ToPlainSelectQuery).
Added a function to convert InsertSelect format to InsertValues format (TryConvertToInsertSelect).
Added a function to convert InsertValues format to InsertSelect format (TryConvertToInsertValues).