Improved the Select clause writing experience
When an IDataRow is passed to the Select method, it is interpreted as selecting all columns in the dataset.
This is equivalent to the behavior of wildcards in SQL.
Carbunql also allows the Select clause to be written multiple times, so it can be written in a similar way even when joining tables.
If a column to be added has already been registered, it will be ignored.
This is because it is common to write columns with higher priority first.