Improved Fluent filtering function
Fixed a bug that caused the filter function to not work if a column had an alias.
Improved the error message when the filter target cannot be found. The name of the table or column that was not found is now displayed in the error message.
The reference to Dapper has been removed from the library itself (it was not necessary).