Enhancement: Carbunql.Fluent namespace
Added the ability to convert a select query to a CTE.
Added the ability to convert a select query to a subquery.
Added the ability to convert a select query to a count query.
Added the ability to convert a select query to an append query.
Added the ability to convert a select query to an update query.
Added the ability to convert a select query to a delete query.
Added the ability to convert a select query to a merge query.
Added the ability to convert a select query to a create table query.