Removed serialization feature
Previously, the clone functionality was implemented using the binary serialization feature of the MessagePack library. However, as the tool's role as an editor specialized for RawSQL has become more defined, the use case for cloning has become less apparent. Given the cost of adding library references and managing attributes for this feature, we have decided to remove it.